refactored components
by splitting them for few new widgets
